+++ docs/changelog.rst
@@ -17,6 +17,7 @@ Changelog
 - Do not show overly verbose raw schemas in Hypothesis output for failed GraphQL tests.
 - The ``schemathesis.graphql.from_url`` loader now uses the usual Schemathesis User-Agent.
 - The Hypothesis database now uses separate entries for each API operation when executed via CLI. It increases its effectiveness when tests are re-run.
+- Module ``schemathesis.loaders`` is moved to ``schemathesis.specs.openapi.loaders``.
 
 **Fixed**